- Change of default telemetry port to 14556 - Added get rssi function to library
